프로토콜 완성 및 구조변경

+++ src/main/java/ase/sensorManager/sensor/Sensor.java
@@ -2,9 +2,9 @@ package ase.sensorManager.sensor;
 
 public class Sensor
 {	
-	public final byte ID;
+	public final int ID;
 	
-	public Sensor(byte id)
+	public Sensor(int id)
 	{
 		this.ID = id;
 	}
